Move tests to AddressUtilTest in correct package, import isHostMatchingPattern statically, import JUnit methods statically to match other tests
